DDP-7443 | extract basic default maker class
